Finish substrate update

+++ relayer/src/relay/client.rs
@@ -105,7 +105,7 @@ impl RuntimeClient for SubstrateRuntimeClient {
             )
             .await?
             .ok_or(anyhow!("Error to get latest beefy block"))?;
-        Ok(latest_beefy_block)
+        Ok(latest_beefy_block as u32)
     }
 
     async fn current_validator_set(&self) -> AnyResult<ValidatorSet> {
@@ -189,7 +189,7 @@ impl RuntimeClient for ParachainRuntimeClient {
             )
             .await?
             .ok_or(anyhow!("Error to get latest beefy block"))?;
-        Ok(latest_beefy_block)
+        Ok(latest_beefy_block as u32)
     }
 
     async fn current_validator_set(&self) -> AnyResult<ValidatorSet> {
